Embeds Harmonic Lines Plot Setup in Frequency Responde Plotter, correct ui file name and adds frequency information on plot
